refactored rewards stuff, added RewardProcessor class
