+++ pytest.ini
@@ -1,1 +1,2 @@
 [pytest]
+asyncio_mode=auto
